feat: refactor api routes
