+++ core/routes/api_config.py
@@ -1,0 +1,2 @@
+LONG_TIMEOUT = 60
+SHORT_TIMEOUT = 60
