+++ core/routes/engine_api.py
@@ -14,7 +14,7 @@ from ninja import Router
 
 from core.schemas import ConnectorSchema, DetailSchema, PackageSchema, PromptSchema, SyncSchema
 
-from .models import (
+from ..models import (
     Connector,
     Package,
     Prompt,
