+++ core/routes/stream_api.py
@@ -8,16 +8,17 @@ Author: Rajashekar Varkala @ valmi.io
 
 import json
 import logging
-
-import requests
-from ninja import Router
-from pydantic import Json
-from decouple import config
-from core.api import SHORT_TIMEOUT
-from core.schemas import GenericJsonSchema
-from .models import ValmiUserIDJitsuApiToken
 from typing import Optional
 
+import requests
+from decouple import config
+from ninja import Router
+from pydantic import Json
+
+from core.routes.api_config import SHORT_TIMEOUT
+from core.schemas import GenericJsonSchema
+
+from ..models import ValmiUserIDJitsuApiToken
 
 router = Router()
